requirements.txt
asgiref==3.10.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
crispy-bootstrap5==2025.6
cryptography==46.0.3
dj-database-url==3.0.1
Django==4.2.25
django-allauth==65.13.1
django-crispy-forms==2.5
idna==3.11
oauthlib==3.3.1
pillow==11.3.0
psycopg2-binary==2.9.11
pycparser==2.23
PyJWT==2.10.1
python-dotenv==1.2.1
requests==2.32.5
requests-oauthlib==2.0.0
sqlparse==0.5.3
typing_extensions==4.15.0
urllib3==2.6.2
whitenoise==6.11.0
